model_codes/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
accelerate=1.1.1=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.11.0=pypi_0
aiosignal=1.3.1=pypi_0
attrs=24.2.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.9.24=haa95532_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
datasets=3.1.0=pypi_0
dill=0.3.8=pypi_0
evaluate=0.4.3=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.54.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.2.0=pypi_0
huggingface-hub=0.26.2=pypi_0
idna=3.10=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.7=pypi_0
libffi=3.4.4=hd77b12b_1
markupsafe=2.1.5=pypi_0
matplotlib=3.9.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
networkx=3.2.1=pypi_0
numpy=1.26.3=pypi_0
openssl=3.0.15=h827c3e9_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=10.2.0=pypi_0
pip=24.2=py311haa95532_0
propcache=0.2.0=pypi_0
protobuf=5.28.3=pypi_0
psutil=6.1.0=pypi_0
pyarrow=18.0.0=pypi_0
pyparsing=3.2.0=pypi_0
python=3.11.10=h4607a30_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=py311haa95532_0
six=1.16.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
tokenizers=0.20.3=pypi_0
torch=2.5.1+cu121=pypi_0
torchaudio=2.5.1+cu121=pypi_0
torchvision=0.20.1+cu121=pypi_0
tqdm=4.67.0=pypi_0
transformers=4.46.2=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
wheel=0.44.0=py311haa95532_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h8cc25b3_1
yarl=1.17.1=pypi_0
zlib=1.2.13=h8cc25b3_1
